# backend/src/OurTube.Domain/Entities/VideoPreview.cs
using OurTube.Domain.Interfaces;

namespace OurTube.Domain.Entities;

public class VideoPreview : Base, IBlob
{
    public Guid VideoId { get; set; }
    public string FileName { get; set; }
    public string Bucket { get; set; }
    
    // Navigation
    public Video Video { get; set; }
}