# docs/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "OurTube API",
    "version": "v1"
  },
  "paths": {
...
    "/users/me/watch-history/videos/{videoId}": {
      "post": {
        "tags": [
          "History"
        ],
        "summary": "Добавить видео в историю просмотров.",
        "parameters": [
          {
            "name": "videoId",
            "in": "path",
            "description": "Идентификатор видео.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Запрос с информацией о просмотренном видео.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostViewsRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PostViewsRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PostViewsRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Видео успешно добавлено в историю."
          },
          "400": {
            "description": "Неверный формат входных данных.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Пользователь не авторизован.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Элемент не найден.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Неизвестная ошибка сервера.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "History"
        ],
        "summary": "Удалить видео из истории просмотров.",
        "parameters": [
          {
            "name": "videoId",
            "in": "path",
            "description": "Идентификатор видео.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Видео успешно удалено из истории."
          },
          "401": {
            "description": "Пользователь не авторизован.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Видео не найдено в истории.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Неизвестная ошибка сервера.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/users/me/watch-history": {
      "delete": {
        "tags": [
          "History"
        ],
        "summary": "Очистить всю историю просмотров пользователя.",
        "responses": {
          "204": {
            "description": "История просмотров успешно очищена."
          },
          "401": {
            "description": "Пользователь не авторизован.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Элемент не найден.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Неизвестная ошибка сервера.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "History"
        ],
        "summary": "Получить историю просмотров пользователя с пагинацией и фильтрацией.",
        "parameters": [
          {
            "name": "SearchQuery",
            "in": "query",
            "description": "Строка поиска по названиям",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Limit",
            "in": "query",
            "description": "Максимальное количество комментариев, возвращаемых за один запрос. (По умолчанию 10)",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "After",
            "in": "query",
            "description": "Смещение для пагинации — номер комментария, после которого следует вернуть следующие.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "Reload",
            "in": "query",
            "description": "Флаг, указывающий, нужно ли перезагрузить комментарии полностью, игнорируя кэш.",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "История просмотров успешно получена.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MinVideoListReply"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MinVideoListReply"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MinVideoListReply"
                }
              }
            }
          },
          "401": {
            "description": "Пользователь не авторизован.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Элемент не найден.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Неизвестная ошибка сервера.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/videos/{videoId}": {
      "get": {
        "tags": [
          "Videos"
        ],
        "summary": "Получить подробную информацию о видео по его идентификатору.",
        "parameters": [
          {
            "name": "videoId",
            "in": "path",
            "description": "Идентификатор видео.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Видео найдено и возвращены его данные.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Video"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Video"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Video"
                }
              }
            }
          },
          "404": {
            "description": "Видео с указанным идентификатором не найдено.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Ошибка сервера.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApplicationUser": {
        "required": [
          "userName"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Уникальный идентификатор пользователя.",
            "format": "uuid"
          },
          "userName": {
            "type": "string",
            "description": "Имя пользователя (отображаемое имя или логин).",
            "nullable": true
          },
          "isSubscribed": {
            "type": "boolean",
            "description": "Флаг, указывающий, подписан ли текущий автор запроса на данного пользователя."
          },
          "subscribersCount": {
            "type": "integer",
            "description": "Количество подписчиков у пользователя.",
            "format": "int32"
          },
          "userAvatar": {
            "$ref": "#/components/schemas/UserAvatar"
          }
        },
        "additionalProperties": false,
        "description": "Модель данных пользователя, возвращаемая в ответах API."
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Машинно-обрабатываемый код ошибки (например, `UserNotFound`, `InvalidEmail`).",
            "nullable": true
          },
          "statusCode": {
            "type": "integer",
            "description": "HTTP-статус код, связанный с ошибкой (например, `400`, `404`, `500`).",
            "format": "int32"
          },
          "message": {
            "type": "string",
            "description": "Человекочитаемое сообщение об ошибке, предназначенное для отображения пользователю.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Модель ошибки, возвращаемая в ответах API."
      },
...
      "MinVideo": {
        "required": [
          "preview",
          "title",
          "user"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Уникальный идентификатор видео.",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "description": "Название видео.",
            "nullable": true
          },
          "viewsCount": {
            "type": "integer",
            "description": "Количество просмотров видео.",
            "format": "int32"
          },
          "vote": {
            "type": "boolean",
            "description": "Голос пользователя за видео: `true` — лайк, `false` — дизлайк,\r\nnull — пользователь не голосовал.",
            "nullable": true
          },
          "duration": {
            "type": "string",
            "description": "Продолжительность видео.",
            "format": "date-span"
          },
          "endTime": {
            "type": "string",
            "description": "Время окончания просмотра видео (если есть), например для отметки прогресса.\r\nМожет быть null.",
            "format": "date-span",
            "nullable": true
          },
          "createdDate": {
            "type": "string",
            "description": "Дата и время создания видео (в формате UTC).",
            "format": "date-time"
          },
          "preview": {
            "$ref": "#/components/schemas/VideoPreview"
          },
          "user": {
            "$ref": "#/components/schemas/ApplicationUser"
          }
        },
        "additionalProperties": false,
        "description": "Минимальная информация о видео для отображения в списках и превью."
      },
      "MinVideoListReply": {
        "required": [
          "elements"
        ],
        "type": "object",
        "properties": {
          "elements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MinVideo"
            },
            "description": "Коллекция элементов текущей страницы.",
            "nullable": true
          },
          "nextAfter": {
            "type": "integer",
            "description": "Параметр, по которому можно запросить следующую страницу данных (например, ID или индекс).",
            "format": "int32",
            "nullable": true
          },
          "hasMore": {
            "type": "boolean",
            "description": "Указывает, есть ли ещё данные после текущей страницы."
          }
        },
        "additionalProperties": false,
        "description": "Универсальная модель для передачи данных с постраничной навигацией (пагинацией)."
      },
...
      "PostViewsRequest": {
        "type": "object",
        "properties": {
          "endTime": {
            "type": "string",
            "description": "Время просмотра видео",
            "format": "date-span"
          }
        },
        "additionalProperties": false,
        "description": "Запрос на регистрацию просмотра"
      },
...
      "UserAvatar": {
        "required": [
          "bucket",
          "fileName",
          "userId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Уникальный идентификатор аватара.",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "description": "Идентификатор пользователя, которому принадлежит аватар.",
            "nullable": true
          },
          "fileName": {
            "type": "string",
            "description": "Имя файла аватара.",
            "nullable": true
          },
          "bucket": {
            "type": "string",
            "description": "Имя хранилища (bucket), где хранится аватар.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Модель, содержащий информацию об аватаре пользователя."
      },
      "Video": {
        "required": [
          "description",
          "files",
          "preview",
          "title",
          "user"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Уникальный идентификатор видео.",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "description": "Заголовок видео.",
            "nullable": true
          },
          "description": {
            "type": "string",
            "description": "Описание видео.",
            "nullable": true
          },
          "likesCount": {
            "type": "integer",
            "description": "Количество лайков.",
            "format": "int32"
          },
          "dislikesCount": {
            "type": "integer",
            "description": "Количество дизлайков.",
            "format": "int32"
          },
          "commentsCount": {
            "type": "integer",
            "description": "Количество комментариев под видео.",
            "format": "int32"
          },
          "viewsCount": {
            "type": "integer",
            "description": "Общее количество просмотров видео.",
            "format": "int32"
          },
          "vote": {
            "type": "boolean",
            "description": "Голос текущего пользователя по видео: `true` — лайк, `false` — дизлайк,\r\nnull — пользователь не голосовал.",
            "nullable": true
          },
          "duration": {
            "type": "string",
            "description": "Продолжительность видео.",
            "format": "date-span"
          },
          "endTime": {
            "type": "string",
            "description": "Время окончания просмотра видео текущим пользователем (если есть).\r\nМожет быть null.",
            "format": "date-span",
            "nullable": true
          },
          "createdDate": {
            "type": "string",
            "description": "Дата и время создания видео (UTC).",
            "format": "date-time"
          },
          "preview": {
            "$ref": "#/components/schemas/VideoPreview"
          },
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VideoPlaylist"
            },
            "description": "Список файлов видео (различные форматы, качества).",
            "nullable": true
          },
          "user": {
            "$ref": "#/components/schemas/ApplicationUser"
          }
        },
        "additionalProperties": false,
        "description": "Модель видео с полной информацией."
      },
      "VideoPlaylist": {
        "required": [
          "bucket",
          "fileName"
        ],
        "type": "object",
        "properties": {
          "resolution": {
            "type": "integer",
            "description": "Разрешение видео (например, 720, 1080).",
            "format": "int32"
          },
          "fileName": {
            "type": "string",
            "description": "Имя файла видео.",
            "nullable": true
          },
          "bucket": {
            "type": "string",
            "description": "Имя хранилища (bucket), где находится видеофайл.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Модель плейлиста M3U8"
      },
      "VideoPreview": {
        "required": [
          "bucket",
          "fileName"
        ],
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string",
            "description": "Имя файла превью изображения.",
            "nullable": true
          },
          "bucket": {
            "type": "string",
            "description": "Имя хранилища (bucket), в котором хранится файл превью.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Модель превью видео с информацией о файле и его расположении."
      }
    }
  },
...
}